pyproject.toml
[tool.poetry]
name = "fitstoolz"
version = "0.0.1b1"
description = "Python libraries for interfacing with FITS data in an intuitive way"
authors = ["Sphesihle Makhathini, Mika Naidoo, Athanaseus Ramaila"]
license = "MIT"
readme = "README.rst"
packages = [
    {include = "fitstoolz"},
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
numpy = "^2.0"
astropy = "^7.1.0"
xarray = "*"
zarr = "*"
dask-ms = "0.2.24"
stimela = "^2.1.2"



#[tool.poetry.scripts]
#imcontsub = "contsub.parser.imcontsub:runit"

[tool.poetry.group.zarr]
optional = true
[tool.poetry.group.zarr.dependencies]
zarr = "*"


[tool.poetry.group.dev]
optional = true
[tool.poetry.group.dev.dependencies]
jupyter = "*"
matplotlib = "*"

[tool.poetry.group.tests]
optional = true
[tool.poetry.group.tests.dependencies]
pytest = "8.4.1"
flake8 = "*"


[tool.poetry.group.docs]
optional = true
[tool.poetry.group.docs.dependencies]
Sphinx = "^8.1"
sphinx-copybutton = "*"
furo = "*"
sphinx-rtd-theme = "*"
sphinx_click = "6.0.0"
sphinx-autodoc-typehints = "*"
docutils = "*"
